Add token `ICellExecutor` providing an entry to customize cell execution
