+++ packages/notebook/src/index.ts
@@ -6,6 +6,7 @@
  */
 
 export * from './actions';
+export * from './cellexecutor';
 export * from './celllist';
 export * from './default-toolbar';
 export * from './executionindicator';
