+++ packages/notebook/src/tokens.ts
@@ -1,14 +1,21 @@
 // Copyright (c) Jupyter Development Team.
 // Distributed under the terms of the Modified BSD License.
 
-import { IWidgetTracker } from '@jupyterlab/apputils';
-import { Cell } from '@jupyterlab/cells';
+import type {
+  ISessionContext,
+  ISessionContextDialogs,
+  IWidgetTracker
+} from '@jupyterlab/apputils';
+import type { Cell } from '@jupyterlab/cells';
+import type { ITranslator } from '@jupyterlab/translation';
 import { Token } from '@lumino/coreutils';
-import { ISignal } from '@lumino/signaling';
-import { Widget } from '@lumino/widgets';
-import { NotebookTools } from './notebooktools';
-import { NotebookPanel } from './panel';
-import { NotebookWidgetFactory } from './widgetfactory';
+import type { ISignal, Signal } from '@lumino/signaling';
+import type { Widget } from '@lumino/widgets';
+import type { NotebookTools } from './notebooktools';
+import type { NotebookPanel } from './panel';
+import type { Notebook } from './widget';
+import type { NotebookWidgetFactory } from './widgetfactory';
+import type { KernelError } from './actions';
 
 /**
  * The notebook widget factory token.
@@ -130,3 +137,62 @@ export interface INotebookTracker extends IWidgetTracker<NotebookPanel> {
    */
   readonly selectionChanged: ISignal<this, void>;
 }
+
+export interface IExecutionOptions {
+  /**
+   * Cell to execute
+   */
+  cell: Cell;
+  /**
+   * Notebook to which the cell belongs
+   */
+  notebook: Notebook;
+  /**
+   * A signal that emits whenever a cell completes execution.
+   */
+  executed: Signal<
+    any,
+    {
+      notebook: Notebook;
+      cell: Cell;
+      success: boolean;
+      error?: KernelError | null;
+    }
+  >;
+  /**
+   * A signal that emits whenever a cell execution is scheduled.
+   */
+  executionScheduled: Signal<any, { notebook: Notebook; cell: Cell }>;
+  /**
+   * Document session context
+   */
+  sessionContext?: ISessionContext;
+  /**
+   * Session dialogs
+   */
+  sessionDialogs?: ISessionContextDialogs;
+  /**
+   * Application translator
+   */
+  translator?: ITranslator;
+}
+
+/**
+ * Notebook cell executor interface
+ */
+export interface ICellExecutor {
+  /**
+   * Execute a cell.
+   *
+   * @param options Cell execution options
+   */
+  runCell(options: IExecutionOptions): Promise<boolean>;
+}
+
+/**
+ * The notebook cell executor token.
+ */
+export const ICellExecutor = new Token<ICellExecutor>(
+  '@jupyterlab/notebook:ICellExecutor',
+  `The notebook cell executor`
+);
